% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{EpsilonExtrapolationVectorOfInverses}
\alias{EpsilonExtrapolationVectorOfInverses}
\title{EpsilonExtrapolationVectorOfInverses
This is a helper function for EpsilonExtrapolation}
\usage{
EpsilonExtrapolationVectorOfInverses(DifferenceMatrix, Method)
}
\arguments{
\item{DifferenceMatrix}{The matrix of the differences in elements to be inverted.}

\item{Method}{"SEA" or "VEA".}
}
\value{
A vector of the result of inverting each (column) vector in a mmatrix.
}
\description{
EpsilonExtrapolationVectorOfInverses
This is a helper function for EpsilonExtrapolation
}
