% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{NicePrint}
\alias{NicePrint}
\title{NicePrint}
\usage{
NicePrint(Number, sigfig = 5)
}
\arguments{
\item{Number}{The number to be formatted.}

\item{sigfig}{The number of significant figures to be taken.}
}
\value{
A nicely formatted string version of the input number for printing to the console.
}
\description{
This is a small function to ensure that decimal places line up for convergence values reported by the fixedpoint function.
}
